// src/Types/Enum.tsx
// empty file

export enum Appointments {
  Unconfirmed = 'Unconfirmed',
  Confirmed = 'Confirmed',
  'Checked-In' = 'Checked-In',
  'Checked Out' = 'Checked Out',
  Rescheduled = 'Rescheduled',
  Canceled = 'Canceled',
  'No-Show' = 'No-Show',
}

export enum AppointmentLabel {
  Unconfirmed = 'Pending Appointment',
  Confirmed = 'confirmed appointment',
  'Checked-In' = 'Check in appointment',
  'Checked Out' = 'Check out',
  Rescheduled = 'reschedule appointment',
  Canceled = 'Canceled appointment',
  'No-Show' = 'No show',
}